Slight cleanup ensureConnectionIsEncrypted
